Adjust integration tests for importing back into same (empty) project.
Combine export and import tests into one suite with sub-tests. Use the
output from export to get the serialized yaml string instead of the
user.yaml test data being written to disk, as this is constantly
changing and triggering git diffs, and is not a long term behaviour we
will keep. (just temporary for debugging as we develop)