Submit backup requests to Ark on archival.
Add additional namespace to exclude from archival.

Drop obsolete archival code now replaced by Ark.

Integration test framework is not useful at this point so it too is
being removed. It will be re-used if we find a way to test the archivist
plus ark in this fashion.

Update code for openshift client-go.