Improve idempotency for Ansible role.
Stop fully reprocessing the template on every running, implicitly
leading to a slow re-build which may not be necessary.

Instead pipe template process output to oc apply, which maintains
previous state and will update them to match the latest requested.

Add a role with associated python module to intelligently check if a
git build is required.